app/requirements.txt
chromadb==0.6.2
Pillow==10.0.0
open-clip-torch==2.31.0
ollama==0.4.7
fastapi==0.110.0
uvicorn==0.29.0
google-genai==1.5.0
python-multipart==0.0.5
opencv-python==4.7.0.72
torch==2.0.1
torchvision==0.15.2
numpy==1.24.2
tqdm==4.65.0
requests==2.28.2
git+https://github.com/facebookresearch/segment-anything.git